- Refactored module/file structure CNN models.
