+++ src/tmap_defectdetector/model/cnn/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
